- Implementado o relatório de resultado de exames.

+++ MyLabSys/Services/OrdemServicoService.cs
@@ -17,6 +17,8 @@ namespace MyLabSys.Services {
         private const string EXAME_FEZES = "Exame de fezes";
         private const string TGO_TGP = "TGO (AST) TGP (ALP)";
         private const string COVID19 = "Teste COVID-19 (RT-PCR)";
+        private const string VLDR = "VLDR";
+        private const string COLESTEROL = "Colesteral";
 
         public OrdemServicoService(MyLabSysContext db) {
             _db = db;
@@ -175,34 +177,34 @@ namespace MyLabSys.Services {
             foreach (var dadosExame in dadosExamesOrdemServico) {
                 switch (dadosExame.Descricao) {
                     case HEMOGRAMA: {
-                        stringBuilder.Append("Homoglobina: 15,1 ---------- 13 a 16 g/dL");
-                        stringBuilder.Append("\nHematrocrito: 44,8 ---------- 38 a 50%");
+                        stringBuilder.Append("Homoglobina: 15,1 ---------- 13 a 16 g/dL ............");
+                        stringBuilder.Append("Hematrocrito: 44,8 ---------- 38 a 50%");
 
                         descricaoResultadoExame = stringBuilder.ToString();
                         stringBuilder.Clear();
                         break;
                     }
                     case GLICEMIA: {
-                        stringBuilder.Append("Amosta em Jejum: 78 mg/dL");
-                        stringBuilder.Append("\nAmosta após 1h: 135 mg/dL");
-                        stringBuilder.Append("\nAmosta apos 2h: 121 mg/dL");
+                        stringBuilder.Append("Amosta em Jejum: 78 mg/dL ............");
+                        stringBuilder.Append("Amosta após 1h: 135 mg/dL ............");
+                        stringBuilder.Append("Amosta apos 2h: 121 mg/dL");
 
                         descricaoResultadoExame = stringBuilder.ToString();
                         stringBuilder.Clear();
                         break;
                     }
                     case EXAME_URINA: {
-                        stringBuilder.Append("Cor: Amarelado");
-                        stringBuilder.Append("\nDensidade: 1,04");
-                        stringBuilder.Append("\nPH: Ácida - 6,5");
+                        stringBuilder.Append("Cor: Amarelado ............");
+                        stringBuilder.Append("Densidade: 1,04 ............");
+                        stringBuilder.Append("PH: Ácida - 6,5 ............");
 
                         descricaoResultadoExame = stringBuilder.ToString();
                         stringBuilder.Clear();
                         break;
                     }
                     case EXAME_FEZES: {
-                        stringBuilder.Append("Ovos de Ascaris Lumbricoides: Não detectado");
-                        stringBuilder.Append("\nLarvas de Strongyloides Tercoralis: Não detectado");
+                        stringBuilder.Append("Ovos de Ascaris Lumbricoides: Não detectado ............");
+                        stringBuilder.Append("Larvas de Strongyloides Tercoralis: Não detectado");
 
                         descricaoResultadoExame = stringBuilder.ToString();
                         stringBuilder.Clear();
@@ -217,6 +219,21 @@ namespace MyLabSys.Services {
                     }
                     case COVID19: {
                         stringBuilder.Append("SARS COV2 RNA: POSITVO");
+
+                        descricaoResultadoExame = stringBuilder.ToString();
+                        stringBuilder.Clear();
+                        break;
+                    }
+                    case VLDR: {
+                        stringBuilder.Append("VLDR (Sífilis): NEGATIVO");
+
+                        descricaoResultadoExame = stringBuilder.ToString();
+                        stringBuilder.Clear();
+                        break;
+                    }
+                    case COLESTEROL: {
+                        stringBuilder.Append("LDH (Ruim): 200 mg/dL");
+                        stringBuilder.Append("HD (Bom): 60 mg/dL");
 
                         descricaoResultadoExame = stringBuilder.ToString();
                         stringBuilder.Clear();
@@ -259,7 +276,12 @@ namespace MyLabSys.Services {
                 .ToArray();
 
             foreach (var resultadoExame in resultadosExames) {
-                _db.Remove(resultadoExame);
+                var existeResultadoExame = resultadoExame != null;
+
+                if (existeResultadoExame) {
+                    _db.Remove(resultadoExame);
+
+                }
             }
         }
